Moved inference loop into testing utils
